Move fieldPosition to its own file
