+++ src/enums/field-position.ts
@@ -1,0 +1,6 @@
+
+export enum FieldPosition {
+  CENTER,
+  LEFT,
+  RIGHT
+}
