+++ src/interfaces/attack-move-result.ts
@@ -1,0 +1,12 @@
+import type { BattlerIndex } from "#app/battle";
+import type { DamageResult } from "#app/field/pokemon";
+import type { Moves } from "#enums/moves";
+
+export interface AttackMoveResult {
+  move: Moves;
+  result: DamageResult;
+  damage: number;
+  critical: boolean;
+  sourceId: number;
+  sourceBattlerIndex: BattlerIndex;
+}
